Add OmniEdge: No Code P2P layer 2 mesh VPN
A rising No Code P2P layer 2 mesh VPN for Enterprise with Zero Config. They found the team by a single tweet and build all native clients, Swift for macOS and iOS, Kotlin for Android, C++ for Windows, and Golang for Cli. A customize supernode is also avaiable for self-hosted to improve speed and security.