Correct name of tunnel.pyjam.as
Hello! 

Thank you so much for adding our project https://tunnel.pyjam.as to your list!

This is just a small correction, but `pyjam.as` is the name of our organization, and the project name is `tunnel.pyjam.as`

I also added a link to our source code

Thank you again! :)