Add bifrost
Bifrost is intended to be a tool for securely tunneling protocols over
unreliable networks.

I placed it under the more advanced tools section since it also is capable of
being used as a library for building apps with Go and is not strictly about
tunneling but rather building dynamically configurable networked Go apps.

Signed-off-by: Christian Stewart <christian@aperture.us>